+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 Added
 *******
 - Implemented :meth:`~sheraf.queryset.QuerySet.__bool__` :issue:`41`
+- Improved :meth:`~sheraf.queryset.QuerySet.__repr__` :issue:`31`
 
 [0.4.19] - 2021-04-23
 =====================